A self implemented IEEE-754 FP.
