+++ Task 1/2019201409于倬浩/test.sh
@@ -1,0 +1,2 @@
+gcc -O2 -fPIC -shared fast.c -o ./my.so -w
+./double-checker ./my.so -f calculate -r 10000000 -o '/'
